Show tags in lazily loaded links
